requirements.txt
adal==1.2.3
amqp==2.5.2
aniso8601==7.0.0
apns2==0.7.1
appdirs==1.4.3
APScheduler==3.6.3
asgiref==3.2.7
asn1crypto==1.3.0
astroid==2.4.1
atomicwrites==1.4.0
attrs==19.3.0
billiard==3.6.3.0
CacheControl==0.12.6
cachetools==4.1.0
CairoSVG==2.4.2
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
coverage==5.1
cryptography==2.9.2
cssselect2==0.3.0
Cython==0.29.17
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.6
django-celery-beat==2.0.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-heroku==0.3.1
django-oauth-toolkit==1.3.2
django-sslserver==0.22
django-timezone-field==4.0
djangorestframework==3.11.0
docutils==0.16
entrypoints==0.3
filelock==3.0.12
firebase-admin==4.2.0
flake8==3.7.9
flake8-pep3101==1.3.0
funcsigs==1.0.2
future==0.18.2
get==2019.4.13
google-api-core==1.17.0
google-api-python-client==1.8.2
google-auth==1.14.2
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-storage==1.28.1
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
googletrans==2.4.0
graphene==2.1.8
graphene-django==2.9.0
graphql-core==2.3.1
graphql-relay==2.0.1
grpcio==1.28.1
gunicorn==20.0.4
h2==3.2.0
hpack==3.0.0
html5lib==1.0.1
http-ece==1.1.0
httpie==2.1.0
httplib2==0.17.3
hyper==0.7.0
hyperframe==5.2.0
idna==2.9
importlib-metadata==1.6.0
isodate==0.6.0
isort==4.3.21
itypes==1.2.0
jeepney==0.4.3
Jinja2==2.11.2
JSON-minify==0.3.0
keyring==21.2.1
kombu==4.6.8
lazy-object-proxy==1.4.3
lockfile==0.12.2
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
msgpack==1.0.0
msrest==0.6.13
msrestazure==0.6.3
mysql-connector-python==8.0.20
oauthlib==3.1.0
olefile==0.46
packaging==20.3
pep8==1.7.1
pew==1.2.0
Pillow==7.1.2
pipenv==2018.11.26
pluggy==0.13.1
post==2019.4.13
promise==2.3
protobuf==3.11.3
psutil==5.7.0
psycopg2==2.7.6.1
psycopg2-binary==2.8.5
public==2019.4.13
py==1.8.1
py-vapid==1.7.0
pyaml==20.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.5.2
pyparsing==2.4.7
Pyphen==0.9.5
pytest==5.4.2
python-crontab==2.4.2
python-daemon==2.2.4
python-dateutil==2.8.1
python-http-client==3.2.7
python-memcached==1.59
python3-openid==3.1.0
pytz==2020.1
pywebpush==1.11.0
PyYAML==5.3.1
query-string==2019.4.13
reportlab==3.5.42
request==2019.4.13
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
Rx==3.1.0
SecretStorage==3.1.2
sendgrid==6.3.0
sh==1.13.1
simplejson==3.17.0
singledispatch==3.4.0.3
six==1.14.0
social-auth-core==3.3.3
sqlparse==0.3.1
tinycss2==1.0.2
toml==0.10.0
tox==3.15.0
typed-ast==1.4.1
tzlocal==2.1
ua-parser==0.10.0
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.9
user-agents==2.1
vine==1.3.0
virtualenv==20.0.20
virtualenv-clone==0.5.4
wcwidth==0.1.9
WeasyPrint==51
webencodings==0.5.1
whitenoise==5.0.1
wrapt==1.12.1
XlsxWriter==1.2.8
zipp==3.1.0
